refactor: Archive callbacks default implementation
